chore: move AntdTypes file into constants
