--- app/renderer/components/AntdTypes.js
@@ -1,26 +1,0 @@
-const BUTTON = {
-  DEFAULT: 'default',
-  PRIMARY: 'primary',
-  DISABLED: 'disabled',
-  DANGER: 'danger',
-};
-
-const ALERT = {
-  ERROR: 'error',
-  WARNING: 'warning',
-  INFO: 'info',
-};
-
-const INPUT = {
-  NUMBER: 'number',
-  TEXT: 'text',
-  TEXTAREA: 'textarea',
-  PASSWORD: 'password',
-  SUBMIT: 'submit',
-};
-
-const ROW = {
-  FLEX: 'flex',
-};
-
-export {BUTTON, ALERT, INPUT, ROW};
