+++ app/renderer/components/Inspector/SelectedElement.js
@@ -10,10 +10,10 @@ import {Alert, Button, Col, Input, Row, Spin, Table, Tooltip} from 'antd';
 import _ from 'lodash';
 import React, {useRef} from 'react';
 
+import {ALERT, ROW} from '../../constants/ANTD_TYPES';
 import {LINKS} from '../../constants/COMMON';
 import {NATIVE_APP} from '../../constants/SESSION_INSPECTOR';
 import {clipboard, shell} from '../../polyfills';
-import {ALERT, ROW} from '../AntdTypes';
 import styles from './Inspector.css';
 
 /**
