+++ app/renderer/components/Session/CapabilityControl.js
@@ -3,8 +3,8 @@ import {Input, Switch} from 'antd';
 import _ from 'lodash';
 import React from 'react';
 
+import {INPUT} from '../../constants/ANTD_TYPES';
 import {log, remote} from '../../polyfills';
-import {INPUT} from '../AntdTypes';
 import SessionStyles from './Session.css';
 
 const getLocalFilePath = async () => {
