+++ app/renderer/components/Session/CloudProviderSelector.js
@@ -2,7 +2,7 @@ import {Button, Col, Modal, Row} from 'antd';
 import _ from 'lodash';
 import React from 'react';
 
-import {BUTTON} from '../AntdTypes';
+import {BUTTON} from '../../constants/ANTD_TYPES';
 import CloudProviders from './CloudProviders';
 import SessionStyles from './Session.css';
