+++ app/renderer/components/Session/ServerTabSauce.js
@@ -1,7 +1,7 @@
 import {Checkbox, Col, Form, Input, Radio, Row} from 'antd';
 import React from 'react';
 
-import {INPUT} from '../AntdTypes';
+import {INPUT} from '../../constants/ANTD_TYPES';
 import SessionStyles from './Session.css';
 
 const sauceUsernamePlaceholder = (t) => {
