+++ app/renderer/constants/ANTD_TYPES.js
@@ -1,0 +1,28 @@
+// Constants used as antd element property values
+
+export const ALERT = {
+  ERROR: 'error',
+  WARNING: 'warning',
+  INFO: 'info',
+};
+
+export const BUTTON = {
+  DEFAULT: 'default',
+  PRIMARY: 'primary',
+  DISABLED: 'disabled',
+  DANGER: 'danger',
+};
+
+export const INPUT = {
+  NUMBER: 'number',
+  TEXT: 'text',
+  TEXTAREA: 'textarea',
+  PASSWORD: 'password',
+  SUBMIT: 'submit',
+};
+
+export const NOTIF = {ERROR: 'error', SUCCESS: 'success'};
+
+export const ROW = {FLEX: 'flex'};
+
+export const TABLE_TAB = {ADD: 'add', REMOVE: 'remove'};
